requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
psutil>=5.8.0
requests>=2.25.0
sqlalchemy>=2.0.0
argparse>=1.4.0
pyyaml>=6.0